+++ packages/ts-components/src/components/olympics/__tests__/OlympicsMedalTable.test.tsx
@@ -4,7 +4,7 @@ import '@testing-library/jest-dom';
 import 'regenerator-runtime';
 import 'jest-styled-components';
 import { OlympicsMedalTable } from '../medal-table/OlympicsMedalTable';
-import { OlympicsKeys } from '../types';
+import { OlympicsKeys } from '../OlympicsKeys';
 
 jest.mock('react-helmet-async', () => ({
   Helmet: 'Helmet',
@@ -23,6 +23,10 @@ describe('<OlympicsMedalTable>', () => {
 
     expect(asFragment()).toMatchSnapshot();
   });
+  it('renders default keys', () => {
+    const { asFragment } = render(<OlympicsMedalTable keys={keys} />);
+    expect(asFragment()).toMatchSnapshot();
+  });
 
   it('renders outside of article', () => {
     const { asFragment } = render(
@@ -34,11 +38,7 @@ describe('<OlympicsMedalTable>', () => {
 
   it('click show all', async () => {
     const { asFragment, getByText, findByText } = render(
-      <OlympicsMedalTable
-        keys={keys}
-        highlighted="AND"
-        sectionColor="sectionColor"
-      />
+      <OlympicsMedalTable keys={keys} highlighted="AND" />
     );
     fireEvent.click(getByText('Show All'));
     await findByText('Collapse');