+++ packages/ts-components/src/components/olympics/__tests__/OlympicsSchedule.test.tsx
@@ -1,11 +1,11 @@
 import React from 'react';
-import { render } from '@testing-library/react';
+import { render, fireEvent } from '@testing-library/react';
 import '@testing-library/jest-dom';
 import 'regenerator-runtime';
 import 'jest-styled-components';
 
 import { OlympicsSchedule } from '../schedule/OlympicsSchedule';
-import { OlympicsKeys } from '../types';
+import { OlympicsKeys } from '../OlympicsKeys';
 
 jest.mock('react-helmet-async', () => ({
   Helmet: 'Helmet',
@@ -23,10 +23,23 @@ describe('<OlympicsSchedule>', () => {
     const { asFragment } = render(<OlympicsSchedule keys={keys} />);
     expect(asFragment()).toMatchSnapshot();
   });
+  it('render default keys', () => {
+    const { asFragment } = render(<OlympicsSchedule />);
+    expect(asFragment()).toMatchSnapshot();
+  });
   it('renders outside of article', () => {
     const { asFragment } = render(
       <OlympicsSchedule keys={keys} inArticle={false} />
     );
+    expect(asFragment()).toMatchSnapshot();
+  });
+  it('click show all', async () => {
+    const { asFragment, getByText, findByText } = render(
+      <OlympicsSchedule keys={keys} />
+    );
+    fireEvent.click(getByText('Show All'));
+    await findByText('Collapse');
+
     expect(asFragment()).toMatchSnapshot();
   });
 });