+++ packages/ts-components/src/components/olympics/schedule/styles.ts
@@ -1,10 +1,12 @@
 import styled from 'styled-components';
 import { breakpoints, colours, fonts } from '@times-components/styleguide';
+import { olympicColour } from '../shared-styles';
 
 export const Container = styled.div<{
-  sectionColour: string;
+  showAll: boolean;
   inArticle: boolean;
 }>`
+  border-top: 2px solid ${olympicColour};
   position: relative;
   margin: 0 auto 20px auto;
 
@@ -31,10 +33,6 @@ export const Container = styled.div<{
       font-size: 16px;
     }
 
-    .pa_UnitListView_ctr {
-      height: 400px;
-    }
-
     .pa_LoadingOverlayContainer_ctr {
       .pa_UnitListView_ctr {
         font-family: ${fonts.supporting};
@@ -43,30 +41,45 @@ export const Container = styled.div<{
         ul.pa_UnitListView_list li {
           background-color: ${colours.functional.backgroundPrimary};
           color: ${colours.functional.brandColour};
+          padding: 10px 6px 8px 10px;
+          &:nth-child(n + 8) {
+            display: ${({ showAll }) => (showAll ? 'block' : 'none')};
+          }
           border-bottom: 1px solid ${colours.functional.keyline};
         }
 
+        .pa_UnitListView_unit {
+          align-items: center;
+        }
+
         .pa_UnitListView_unit-time {
-          color: ${({ sectionColour }) => sectionColour};
+          color: ${olympicColour};
           line-height: 30px;
         }
 
         .pa_UnitListView_unit-text {
           font-family: ${fonts.supporting};
-          font-size: 14px;
+          font-size: 16px;
         }
 
         .pa_UnitListView_discipline {
           font-family: ${fonts.headline};
           text-transform: capitalize;
           font-weight: 400;
+          margin-top: 4px;
         }
+      }
+      .pa_UnitListView_status-live {
+        padding: 8px 20px 4px;
+        align-self: center;
       }
     }
 
     .pa_OdfFooter_ctr {
       font-family: ${fonts.supporting};
       font-size: 12px;
+      top: 60px;
+      position: relative;
     }
 
     .pa_UnitListView_message {
@@ -76,10 +89,21 @@ export const Container = styled.div<{
 
     .pa_Schedule_ctr {
       background: ${colours.functional.backgroundSecondary};
+      padding-bottom: 60px;
     }
 
     .pa_ErrorMessage_ctr {
       background: #ededed;
+    }
+  }
+  .buttonContainer {
+    text-align: center;
+    height: 0;
+    button {
+      background-color: ${colours.functional.backgroundPrimary};
+      :hover {
+        background-color: ${colours.functional.backgroundSecondary};
+      }
     }
   }
 `;